# api/openapi.yaml
openapi: 3.1.0
info:
  title: ZTE C300 Monitoring API
  description: |
    Read-only SNMP collector for ZTE C300 OLTs (C320-derived mappings remain
    in code but are not a separate compatibility guarantee). Live routes
    return PON inventory, ONU detail, uplink topology, Timescale history,
    status/UNI events and unconfigured-ONU discovery. HTTP POST/DELETE only
    touch the collector Redis cache.

    **Multi-OLT:** when the instance is configured with multiple OLTs (the `OLTS`
    env var), every `/api/v1/board/...` and `/api/v1/history/...` path below is
    also available under `/api/v1/olt/{olt_id}/...`, where `{olt_id}` selects
    the target OLT. The configured default OLT continues to answer the bare
    `/api/v1/board/...` and `/api/v1/history/...` paths shown here. `board_id`
    is the physical GPON slot and `pon_id` is validated against that slot's card
    (GTGO=8 or GTGH=16 PON ports). History routes require TimescaleDB.

    API contract version is `v1` (endpoint prefix `/api/v1`). The binary
    version and contract version are exposed via `X-App-Version` and
    `X-API-Version` response headers on every request.
  version: 0.1.0-dev
  contact:
    name: API Support
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8081
    description: Local development server
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/board/{board_id}/pon/{pon_id}:
    get:
      summary: Get all ONUs for a board/PON combination
      description: Retrieves a list of all ONUs connected to the specified board and PON port.
      operationId: getOnusByBoardPon
      tags:
        - ONU
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
      responses:
        "200":
          description: List of ONUs retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ONUInfoPerBoard"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/cache/clear:
    delete:
      summary: Clear cache for a board/PON combination
      description: Removes the Redis cache for the specified board and PON port. Never sends SNMP writes and never deletes an ONU.
      operationId: clearCacheByBoardPon
      tags:
        - Cache
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
      responses:
        "200":
          description: Cache cleared successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Cache deleted successfully
                          board_id:
                            type: integer
                            example: 1
                          pon_id:
                            type: integer
                            example: 1
                        required:
                          - message
                          - board_id
                          - pon_id
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/onu/{onu_id}:
    get:
      summary: Get specific ONU detail
      description: Retrieves detailed information for a specific ONU including optical power levels, uptime, and offline history.
      operationId: getOnuDetail
      tags:
        - ONU
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
        - $ref: "#/components/parameters/OnuId"
      responses:
        "200":
          description: ONU detail retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ONUCustomerInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/onu_id/empty:
    get:
      summary: Get available/empty ONU IDs
      description: Returns a list of ONU IDs that are not currently assigned on the specified board and PON port.
      operationId: getEmptyOnuIds
      tags:
        - ONU ID
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
      responses:
        "200":
          description: List of empty ONU IDs retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/OnuID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/onu_id_sn:
    get:
      summary: Get ONU IDs with serial numbers
      description: Returns a list of ONU IDs along with their corresponding serial numbers for the specified board and PON port.
      operationId: getOnuIdSerialNumbers
      tags:
        - ONU ID
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
        - $ref: "#/components/parameters/NoCache"
      responses:
        "200":
          description: ONU IDs with serial numbers retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/OnuSerialNumber"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/onu_id/update:
    post:
      summary: Force refresh empty ONU ID cache
      description: Triggers a cache refresh for empty ONU IDs on the specified board and PON port.
      operationId: updateEmptyOnuIdCache
      tags:
        - Cache
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
      responses:
        "200":
          description: Cache updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: string
                        example: Success Update Empty ONU_ID
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/paginate/board/{board_id}/pon/{pon_id}:
    get:
      summary: Get paginated ONU list
      description: Retrieves a paginated list of ONUs for the specified board and PON port with metadata about total rows and page count.
      operationId: getOnusPaginated
      tags:
        - ONU
      parameters:
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
        - name: page
          in: query
          description: Page number (1-indexed)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        "200":
          description: Paginated list of ONUs retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ONUInfoPerBoard"
                      meta:
                        $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/board/{board_id}/pon/{pon_id}/onu/{onu_id}/cache/clear:
    delete:
      summary: Invalidate cached ONU and related PON records
      description: Changes Redis cache only; never sends an SNMP write.
      operationId: invalidateOnuCache
      tags: [Cache]
      parameters:
        - $ref: '#/components/parameters/BoardId'
        - $ref: '#/components/parameters/PonId'
        - $ref: '#/components/parameters/OnuId'
      responses:
        '200':
          description: Related cache invalidated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ───────────────────────────────────────────────────────────────────────────
  # Multi-OLT variants. Identical semantics to the bare /api/v1/board/... paths
  # above, but {olt_id} selects which OLT in the OLTS/OLTS_FILE registry to
  # target. The configured DEFAULT_OLT also answers the bare paths above.
  # ───────────────────────────────────────────────────────────────────────────
  /api/v1/uplinks:
    get:
      summary: Auto-detect OLT cards and uplink ethernet ports (default OLT)
      description: >
        SNMP auto-detect of the OLT's physical cards (ENTITY-MIB
        entPhysicalDescr/entPhysicalClass, classified as gpon / control /
        uplink / power) and uplink ethernet ports (IF-MIB ifName/admin/oper/
        speed; xgei_ = 10G, gei_ = 1G, SmartGroupN = LAG). LAG members come
        from ifStackStatus when the agent exposes it. Detection-only — no
        configuration writes. Firmware naming and card-slot interpretation
        require hardware verification. Card slot prefers the reported
        parent-relative position, with an explicit heuristic fallback.
      operationId: getUplinkTopology
      tags:
        - Uplink
      responses:
        "200":
          description: Auto-detected card + uplink-port topology
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UplinkTopology"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/olt/{olt_id}/board/{board_id}/pon/{pon_id}:
    get:
      summary: Get all ONUs for a board/PON (multi-OLT)
      description: Same as the bare path but scoped to the OLT named by `olt_id`.
      operationId: getOnusByBoardPonForOlt
      tags:
        - ONU
      parameters:
        - $ref: "#/components/parameters/OltId"
        - $ref: "#/components/parameters/BoardId"
        - $ref: "#/components/parameters/PonId"
      responses:
        "200":
          description: List of ONUs retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ONUInfoPerBoard"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/olt/{olt_id}/uplinks:
    get:
      summary: Auto-detect OLT cards and uplink ethernet ports (specific OLT)
      description: Multi-OLT variant of /api/v1/uplinks for the OLT selected by olt_id.
      operationId: getUplinkTopologyByOlt
      tags:
        - Uplink
      parameters:
        - $ref: "#/components/parameters/OltId"
      responses:
        "200":
          description: Auto-detected card + uplink-port topology
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/WebResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UplinkTopology"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/history/samples:
    get:
      summary: List stored PON-list samples
      description: |
        Returns TimescaleDB snapshots written by the read-only poller (status,
        RX/TX, serial, position, Ethernet UNI summary). Rows also include
        `status_changed_at`, `previous_status` and `expected_eth_ports` from
        the current identity table. Does not query the OLT. Requires TimescaleDB.

        Also available at `/api/v1/olt/{olt_id}/history/samples`.

        When `count_by` is set, `data` is an aggregate (`group_by`, `run_id`,
        `from`, `to`, `total`, `counts[]`) instead of a sample array.
      operationId: listHistorySamples
      tags: [History]
      parameters:
        - name: run
          in: query
          description: >
            `latest`/`last` selects the newest collection run's time window, or
            pass a numeric run id from `/history/runs`. Overrides `from`/`to`.
          schema: { type: string }
        - name: status
          in: query
          description: Case-insensitive ONU status (Online, Offline, …)
          schema: { type: string }
        - name: serial
          in: query
          schema: { type: string }
        - name: board
          in: query
          schema: { type: integer }
        - name: pon
          in: query
          schema: { type: integer }
        - name: onu_id
          in: query
          schema: { type: integer }
        - name: from
          in: query
          description: RFC3339 start (default now-24h). Ignored when `run` is set.
          schema: { type: string, format: date-time }
        - name: to
          in: query
          description: RFC3339 end (default now). Ignored when `run` is set.
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          description: Max rows (default 500, or 2000 when `run` is set). Unused with `count_by`.
          schema: { type: integer, default: 500, maximum: 2000 }
        - name: count_by
          in: query
          description: Return counts grouped by this field instead of sample rows.
          schema:
            type: string
            enum: [status, board, pon, onu_type, eth_link]
      responses:
        '200':
          description: Sample list or count_by aggregate
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/HistorySamplesResponse'
                  - $ref: '#/components/schemas/HistoryCountsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/history/runs:
    get:
      summary: List poller collection runs
      description: Duration, PON success/error counts and ONU rows written.
      operationId: listCollectionRuns
      tags: [History]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: Collection run list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryRunsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/history/unauth:
    get:
      summary: List unconfigured ONUs last seen by the OLT
      description: >
        Last discovery snapshot. `unsupported` means the firmware walk is not
        confirmed. Does not delete or rewrite onu_samples.
        Also available at `/api/v1/olt/{olt_id}/history/unauth`.
      operationId: listUnauth
      tags: [History]
      responses:
        '200':
          description: Unconfigured-ONU discovery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryUnauthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        API key for /api/v1 routes. With a single `API_KEY` the key only gates
        access. With the per-tenant `API_USERS` registry, the key also identifies
        the caller (user_id + role): a user sees only the OLTs whose `user_id`
        matches theirs — requesting another tenant's OLT returns 404 (not 403, to
        avoid enumeration) — while an `admin` role sees all OLTs. A missing or
        unknown key returns 401. Startup requires API_KEY or API_USERS.
...